错误sql修改

+++ globalshop-biz1-app/src/main/java/com/wangqin/globalshop/biz1/app/dal/dataObject/AuthOrganizationDO.java
@@ -17,7 +17,7 @@ public class AuthOrganizationDO {
 
     private Long pid;
 
-    private Byte seq;
+    private Integer seq;
 
     private Date gmtCreate;
 
@@ -85,11 +85,11 @@ public class AuthOrganizationDO {
         this.pid = pid;
     }
 
-    public Byte getSeq() {
+    public Integer getSeq() {
         return seq;
     }
 
-    public void setSeq(Byte seq) {
+    public void setSeq(Integer seq) {
         this.seq = seq;
     }
